Moved telemetry prefs to AppearanceTab
