+++ src/main/java/org/jabref/gui/preferences/appearance/AppearanceTab.fxml
@@ -69,4 +69,6 @@
     <CheckBox fx:id="inspectionWarningDuplicate"
               text="%Warn about unresolved duplicates when closing inspection window"/>
     <CheckBox fx:id="confirmDelete" text="%Show confirmation dialog when deleting entries"/>
+
+    <CheckBox fx:id="collectTelemetry" text="%Collect and share telemetry data to help improve JabRef"/>
 </fx:root>
