+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.fxml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.VBox?>
 <fx:root spacing="10.0" type="VBox"
@@ -8,6 +7,6 @@
          fx:controller="org.jabref.gui.preferences.library.LibraryTab">
     <Label styleClass="titleHeader" text="%General"/>
 
-    <CheckBox fx:id="collectTelemetry" text="%Collect and share telemetry data to help improve JabRef"/>
+
 
 </fx:root>
