+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.java
@@ -1,8 +1,5 @@
 package org.jabref.gui.preferences.library;
 
-import javafx.fxml.FXML;
-import javafx.scene.control.CheckBox;
-
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
 import org.jabref.logic.l10n.Localization;
@@ -11,7 +8,7 @@ import com.airhacks.afterburner.views.ViewLoader;
 
 public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> implements PreferencesTab {
 
-    @FXML private CheckBox collectTelemetry;
+
 
     public LibraryTab() {
         ViewLoader.view(this)
@@ -26,7 +23,5 @@ public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> i
 
     public void initialize() {
         this.viewModel = new LibraryTabViewModel(dialogService, preferencesService.getGeneralPreferences(), preferencesService.getTelemetryPreferences());
-
-        collectTelemetry.selectedProperty().bindBidirectional(viewModel.collectTelemetryProperty());
     }
 }
